Rename throttle env var to API_RATE_LIMIT & move throttle util to bootstrap/util
