+++ blockcypher/src/config.ts
@@ -4,7 +4,7 @@ import { util } from '@chainlink/ea-bootstrap'
 
 export const DEFAULT_ENDPOINT = 'balance'
 
-export const ENV_REQUEST_THROTTLE = 'API_RATE_LIMIT'
+export const ENV_RATE_LIMIT = 'API_RATE_LIMIT'
 
 export type Config = types.Config & {
   ratelimit: number
@@ -12,7 +12,7 @@ export type Config = types.Config & {
 
 export const makeConfig = (prefix = ''): Config => {
   const config = Requester.getDefaultConfig(prefix)
-  const throttle = util.getEnv(ENV_REQUEST_THROTTLE, prefix)
-  if (throttle) config.throttle = Number(throttle)
+  const ratelimit = util.getEnv(ENV_RATE_LIMIT, prefix)
+  if (ratelimit) config.ratelimit = Number(ratelimit)
   return config
 }
