+++ blockcypher/src/endpoint/balance.ts
@@ -3,6 +3,7 @@ import { balance } from '@chainlink/ea-factories'
 import { Account } from '@chainlink/types'
 import { Config } from '../config'
 import { CoinType, ChainType, isCoinType, isChainType } from '.'
+import { util } from '@chainlink/ea-bootstrap'
 
 export const Name = 'balance'
 
@@ -29,47 +30,6 @@ const getChainId = (coin: CoinType, chain: ChainType): string => {
   }
 }
 
-/**
- * Chunk an array into a nested array
- *
- * @param amount number the max size of the chunks
- * @param data array of arbitrary data
- *
- * @returns 2d array
- */
-const chunk = (amount: number, data: any[]) => {
-  const output: any[][] = []
-  const length = data.length
-  if (amount < 1 || length < 1) return output
-  const chunks = Math.ceil(data.length / amount)
-  for (let i = 0; i < chunks; i++) {
-    const offset = amount * i
-    const slice = data.slice(offset, offset + amount)
-    output.push(slice)
-  }
-  return output
-}
-
-/**
- * Delays the amount of requests sent per second
- *
- * @param amount maximum number of requests per second
- * @param data array of arbitrary data
- * @param exec function handler of a chunk
- *
- * @returns array of response data
- */
-const rateLimit = async (amount: number, data: any[], exec: any) => {
-  const chunks = chunk(amount, data)
-  const delay = 1000
-  const withDelay = async (c: any, i: number) => {
-    await new Promise((resolve) => setTimeout(resolve, i * delay))
-    return await exec(c)
-  }
-  const output = await Promise.all(chunks.map(withDelay))
-  return output.flat()
-}
-
 const getBalances: balance.GetBalances<Config> = async (accounts, config) => {
   const addresses = accounts.map((a) => a.address)
   const { coin, chain } = accounts[0]
@@ -90,11 +50,11 @@ const getBalances: balance.GetBalances<Config> = async (accounts, config) => {
     })
 
   const response = config.ratelimit
-    ? await rateLimit(config.ratelimit, addresses, _getAddrBal)
+    ? await util.throttle(config.ratelimit, addresses, _getAddrBal)
     : await _getAddrBal(addresses)
 
   const addrLookup: { [key: string]: AddressBalance } = {}
-  response.forEach((r) => (addrLookup[r.address] = r))
+  response.forEach((r: any) => (addrLookup[r.address] = r))
 
   const addBalance = (acc: Account) => ({
     ...acc,
